refactor(login-google): export queries
